Norme

+++ minitalk_client.c
@@ -11,14 +11,31 @@
 /* ************************************************************************** */
 
 #include <signal.h>
+#include <unistd.h>
 #include "libft/libft.h"
 
 int	main(int argc, char **argv)
 {
 	int	pid;
+	int	i;
+	int	j;
 
-	(void)argc;
+	if (argc != 3)
+		return (0);
 	pid = ft_atoi(argv[1]);
-	kill(pid, SIGUSR1);
+	i = 0;
+	while (argv[2][i])
+	{
+		j = 0;
+		while (j < argv[2][i])
+		{
+			kill(pid, SIGUSR1);
+			usleep(5);
+			j++;
+		}
+		kill(pid, SIGUSR2);
+		usleep(5);
+		i++;
+	}
 	return (0);
 }
